Working on processing SPSS / Stata files
